fix bugs and rename Components
