--- src/components/AboutFormModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AboutFormModal';
