+++ src/components/ProjectDetails/ProjectDetailsTypes.ts
@@ -1,4 +1,4 @@
-export interface IAboutFormProps {
+export interface IProjectDitealsProps {
   btnTitle: string;
   formTitle: string;
   checkBoxShow: boolean;
