+++ src/components/ProjectDetailsModal/ProjectDetailsModal.module.scss
