+++ src/components/ProjectDetailsModal/ProjectDetailsModal.tsx
@@ -1,17 +1,23 @@
-import { FC } from 'react';
-import { IAboutFormModalProps } from './AboutFormModalTypes';
+import { FC, useEffect } from 'react';
+import { IProjectDitealsProps } from './ProjectDetailsModalTypes';
 import GetProjectForm from '../GetProjectForm/GetProjectForm';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faX } from '@fortawesome/free-solid-svg-icons';
-import styles from './AboutFormModal.module.scss';
+import styles from './ProjectDetailsModal.module.scss';
 
-const AboutFormModal: FC<IAboutFormModalProps> = ({
-  showFormModal,
+const ProjectDetailsModal: FC<IProjectDitealsProps> = ({
   setShowFormModalBlock,
   checkBoxShow,
   formTitle,
 }) => {
-  return showFormModal ? (
+  useEffect(() => {
+    const body = document.body;
+    body.style.overflow = 'hidden';
+    return () => {
+      body.style.overflow = 'visible';
+    };
+  }, []);
+  return (
     <div className={styles.lightbox_inner}>
       <div className={styles.lightbox_content}>
         <div className={styles.block_form}>
@@ -28,7 +34,7 @@ const AboutFormModal: FC<IAboutFormModalProps> = ({
         </div>
       </div>
     </div>
-  ) : null;
+  );
 };
 
-export default AboutFormModal;
+export default ProjectDetailsModal;
