+++ src/components/ProjectDetailsModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProjectDetailsModal';
