+++ tsconfig.json
@@ -23,5 +23,5 @@
   },
   "include": [
     "src"
-  ]
+, "src/components/ProjectDetailsModal/ProjectDetailsModal.tsx"  ]
 }
